# openapi.yml
openapi: 3.0.0
info:
    title: Health Checker API
    version: 1.0.1
    description: API for registering and monitoring URL health checks

servers:
    - url: http://localhost:8080
      description: Local development server

paths:
    /health:
        get:
            summary: Get the health status of the API
            operationId: getHealth
            responses:
                "204":
                    description: API is healthy
                "500":
                    description: API is unhealthy
                    $ref: "#/components/responses/InternalServerError"
    /register:
        post:
            summary: Register a new URL for health checking
            operationId: registerTarget
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/Target"
            responses:
                "204":
                    description: Target successfully registered
                "400":
                    $ref: "#/components/responses/BadRequest"
                "405":
                    $ref: "#/components/responses/NotAllowed"
                "500":
                    $ref: "#/components/responses/InternalServerError"

    /unregister/{id}:
        delete:
            summary: Unregister a URL from health checking
            operationId: unregisterTarget
            parameters:
                - name: id
                  in: path
                  required: true
                  description: The unique identifier of the target to unregister
                  schema:
                      type: string
            responses:
                "204":
                    description: Target successfully unregistered
                "400":
                    $ref: "#/components/responses/BadRequest"
                "404":
                    $ref: "#/components/responses/NotFound"
                "405":
                    $ref: "#/components/responses/NotAllowed"
                "500":
                    $ref: "#/components/responses/InternalServerError"

    /status:
        get:
            summary: Get health check status for all registered targets
            operationId: getStatus
            responses:
                "200":
                    description: List of health check results
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/HealthCheckResult"
                "405":
                    $ref: "#/components/responses/NotAllowed"

components:
    responses:
        BadRequest:
            description: Bad request
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/Error"
        NotFound:
            description: Resource not found
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/Error"
        NotAllowed:
            description: Method not allowed
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/Error"
        InternalServerError:
            description: Internal server error
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/Error"
    schemas:
        Target:
            type: object
            required:
                - url
                - id
            properties:
                url:
                    type: string
                    format: uri
                    description: The URL to be monitored
                id:
                    type: string
                    description: Unique identifier for the target

        HealthCheckResult:
            type: object
            required:
                - id
                - url
                - status
                - healthy
                - timestamp
                - duration_seconds
            properties:
                id:
                    type: string
                    description: Target identifier
                url:
                    type: string
                    format: uri
                    description: The monitored URL
                status:
                    type: integer
                    description: HTTP status code from the health check
                healthy:
                    type: boolean
                    description: Whether the target is considered healthy
                timestamp:
                    type: string
                    format: date-time
                    description: When the health check was performed
                duration_seconds:
                    type: number
                    format: float
                    description: Duration of the health check in seconds
                error:
                    type: string
                    description: Error message if the health check failed

        Error:
            type: object
            required:
                - code
                - message
            properties:
                code:
                    type: string
                    description: Error code static for the error type
                message:
                    type: string
                    description: Error message with further details